% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMGMR.R
\name{bayes_inference}
\alias{bayes_inference}
\title{Summary statistics from a posterior distribution}
\usage{
bayes_inference(x, alp)
}
\arguments{
\item{x}{a sample from the posterior which to aggregate from.}

\item{alp}{the confidence level used in computation of the summary statistics}
}
\value{
a matrix with summary statistics
}
\description{
Summary statistics from a posterior distribution
}
