% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMGMR.R
\name{BMGMR}
\alias{BMGMR}
\title{Interface for the BMGMR class}
\usage{
BMGMR(X, U, N, burn_in, likelihood, prior, algorithm_version, d = NULL)
}
\arguments{
\item{X}{observation matrix of dimension p x n}

\item{U}{block matrix with uncertainties of size pn x pn}

\item{N}{number of samples to simulate from the markov chain}

\item{burn_in}{the number of burn-in samples}

\item{likelihood}{the likelihood to use. It currently supports "normal" and
"t".}

\item{prior}{the prior to use. It currently supports "reference" and
"jeffrey".}

\item{algorithm_version}{One of "A" or "B". Both algorithms samples the same
quantities.}

\item{d}{the degrees of freedom which is used when the "t" option is used for
the likelihood.}
}
\value{
a BMGMR class.
}
\description{
This is the main interface to the MCMC algorithms implemented in this
package. The data is on wide format with p rows (variables) and n columns
(observations).
}
