% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_sample_post.R
\name{duplication_matrix}
\alias{duplication_matrix}
\title{A Duplication matrix}
\usage{
duplication_matrix(x)
}
\arguments{
\item{x}{the dimension of the matrix}
}
\value{
a matrix of size x^2
}
\description{
This function creates a duplication matrix of size x
}
