% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMGMR.R
\name{plot.BMGMR}
\alias{plot.BMGMR}
\title{Plot a BMGMR object}
\usage{
\method{plot}{BMGMR}(x, ...)
}
\arguments{
\item{x}{a BMGMR object}

\item{...}{optional arguments to ci2d (if gplots is installed)}
}
\description{
Plot a BMGMR object
}
