% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_sample_post.R
\name{sample_post_nor_jef_marg_mu}
\alias{sample_post_nor_jef_marg_mu}
\title{Metropolis hasting algorithm A}
\usage{
sample_post_nor_jef_marg_mu(X, U, Np)
}
\arguments{
\item{X}{A p x n matrix which contains the observations}

\item{U}{A p n x p n block matrix which contains uncertainties.}

\item{Np}{the number of simulations to perform.}
}
\value{
list with samples from the two marginal distributions
}
\description{
This function implements the algorithm A using the normal likelihood and
jeffreys prior. The number of observations is n and the number of variables
is p.
}
