% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_sample_post.R
\name{sample_post_nor_ref_marg_mu}
\alias{sample_post_nor_ref_marg_mu}
\title{Metropolis Hasting Algorithm A, normal-distribution}
\usage{
sample_post_nor_ref_marg_mu(X, U, Np)
}
\arguments{
\item{X}{A p x n matrix which contains the observations}

\item{U}{A p n x p n block matrix which contains uncertainties.}

\item{Np}{the number of simulations to perform.}
}
\value{
list with samples from the two marginal distributions
}
\description{
A MH algorithm that samples using a likelihood from the normal distribution
and the reference prior.
}
