% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_sample_post.R
\name{sample_post_t_ref_marg_Psi}
\alias{sample_post_t_ref_marg_Psi}
\title{Metropolis Hasting Algorithm B, t-reference}
\usage{
sample_post_t_ref_marg_Psi(X, U, d, Np)
}
\arguments{
\item{X}{A p x n matrix which contains the observations}

\item{U}{A p n x p n block matrix which contains uncertainties.}

\item{d}{the degrees of freedom for the t-distribution}

\item{Np}{the number of simulations to perform.}
}
\value{
list with samples from the two marginal distributions
}
\description{
A MH algorithm that samples using a likelihood from the t distribution
and the reference prior.
}
