% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMGMR.R
\name{summary.BMGMR}
\alias{summary.BMGMR}
\title{Summary statistics from the posterior of a BMGMR class}
\usage{
\method{summary}{BMGMR}(object, alpha = 0.95, ...)
}
\arguments{
\item{object}{is a BMGMR class}

\item{alpha}{is the confidence level used in construction of the credible
intervals}

\item{...}{not used}
}
\value{
a list with summary statistics
}
\description{
It will use the data without burn in
}
